File Descriptions:
Generate_Data.py:
This Python file calls the OpenWeatheMap API, stores the responses as JSON, extracts specific data (city name, date, and temperature) and stores it all in a data frame. 

pipeline_to_database.py:
This python file establishes a connection to the postgresql database. It also takes the data frame created in Generate_data.py, and inserts the elements to a table in the database.